requirements.txt
selenium
Flask
Flask-SQLAlchemy
requests
psycopg2-binary
webdriver_manager.chrome